DSP: More work on dsptool. Minor bugfixes. Add some testdata for dsptool.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2993 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DSPCore/Src/DSPInterpreter.cpp
@@ -1214,7 +1214,7 @@ void addis(const UDSPInstruction& opc)
 {
 	u8 areg = (opc.hex >> 8) & 0x1;
 
-	s64 Imm = (s8)opc.hex;
+	s64 Imm = (s8)(u8)opc.hex;
 	Imm <<= 16;
 	s64 acc = dsp_get_long_acc(areg);
 	acc += Imm;
